+++ src/resources/stores.jl
@@ -10,6 +10,16 @@ struct StoreGetKey{T<:Number} <: ResourceKey
   priority :: T
 end
 
+"""
+    Store{T}(env::Environment; capacity::UInt=typemax(UInt))
+
+A store is a resource that can hold a number of items of type `T`. It is similar to a `Base.Channel` with a finite capacity ([`put!`](@ref) blocks after reaching capacity).
+The [`put!`](@ref) and [`take!`](@ref) functions are a convenient way to interact with such a "channel" in a way mostly compatible with other discrete event and concurrency frameworks.
+
+See [`Container`](@ref) for a more lock-like resource.
+
+Think of `Resource` and `Container` as locks and of `Store` as channels. They block only if empty (on taking) or full (on storing).
+"""
 mutable struct Store{N, T<:Number} <: AbstractResource
   env :: Environment
   capacity :: UInt
@@ -25,8 +35,13 @@ end
 
 function Store{N}(env::Environment; capacity::UInt=typemax(UInt)) where {N}
   Store{N, Int}(env; capacity)
-end
+end    
+    
+"""
+    put!(sto::Store, item::T)
 
+Put an item into the store. Returns the put event, blocking if the store is full.
+"""
 function put(sto::Store{N, T}, item::N; priority::T=zero(T)) where {N, T<:Number}
   put_ev = Put(sto.env)
   sto.put_queue[put_ev] = StorePutKey{N, T}(sto.seid+=one(UInt), item, priority)
@@ -69,3 +84,49 @@ function do_get(sto::Store{N, T}, get_ev::Get, key::StoreGetKey{T}) where {N, T<
   end
   true
 end
+
+"""
+    isready(::Store)
+
+Returns `true` if the store is not empty, similarly to the meaning of `isready` for `Base.Channel`.
+
+```jldoctest
+julia> sim = Simulation(); store = Store{Symbol}(sim); isready(store)
+false
+
+julia> put!(store, :message); isready(store)
+true
+```
+"""
+isready(sto::Store) = sto.load > 0
+
+"""
+    islocked(::Store)
+
+Returns `true` if the store is full, similarly to the meaning of `islocked` for `Base.ReentrantLock`.
+
+```jldoctest
+julia> sim = Simulation(); store = Store{Symbol}(sim; capacity=2); islocked(store)
+false
+
+julia> put!(store, :message); islocked(store)
+false
+
+julia> put!(store, :another_message); islocked(store)
+true
+```
+"""
+islocked(sto::Store) = sto.load==sto.capacity
+
+unlock(::Store) = error("There is no well defined way to \"unlock\" a Store without taking an element out of it. Instead of attempting `unlock` consider using `take!(::Store)` or use a `Resource` instead of a `Store`. Think of `Resource` and `Container` as locks and of `Store` as channels. They block only if empty (on taking) or full (on storing).")
+lock(::Store) = error("There is no well defined way to \"lock\" a Store without storing an element in it. Instead of attempting `lock` consider using `put!(::Store, ...)` or use a `Resource` instead of a `Store`. Think of `Resource` and `Container` as locks and of `Store` as channels. They block only if empty (on taking) or full (on storing).")
+trylock(::Store) = error("There is no well defined way to \"lock\" a Store without storing an element in it. Instead of attempting `lock` consider using `put!(::Store, ...)` or use a `Resource` instead of a `Store`. Think of `Resource` and `Container` as locks and of `Store` as channels. They block only if empty (on taking) or full (on storing).")
+request(::Store) = error("There is no well defined way to \"request\" a Store without storing an element in it. Instead of attempting `request` consider using `put!(::Store, ...)` or use a `Resource` instead of a `Store`. Think of `Resource` and `Container` as locks and of `Store` as channels. They block only if empty (on taking) or full (on storing).")
+tryrequest(::Store) = error("There is no well defined way to \"request\" a Store without storing an element in it. Instead of attempting `request` consider using `put!(::Store, ...)` or use a `Resource` instead of a `Store`. Think of `Resource` and `Container` as locks and of `Store` as channels. They block only if empty (on taking) or full (on storing).")
+
+"""
+    take!(::Store)
+
+An alias for `get(::Store)` for easier interoperability with the `Base.Channel` interface. Blocks if the store is empty.
+"""
+take!(sto::Store, filter::Function=get_any_item; priority::Int=0) = get(sto, filter; priority)